fix(epic-44): resolve styled-jsx server component errors
- Add 'use client' directive to homepage page.tsx
- Extract ListingDetailContent client component from listing detail page
- Refactor [slug]/page.tsx to use client component while keeping SSR data fetching
- Keep generateMetadata for SEO in server component
- Fix import ordering per Biome rules
