+++ frontend/apps/mobile/package.json
@@ -22,5 +22,5 @@
     "@types/react": "~18.2.0",
     "typescript": "^5.3.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
