+++ frontend/apps/ppt-web/package.json
@@ -26,5 +26,5 @@
     "typescript": "^5.3.0",
     "vite": "^5.0.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
