+++ frontend/packages/api-client/package.json
@@ -18,5 +18,5 @@
     "@hey-api/openapi-ts": "^0.36.0",
     "typescript": "^5.3.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
