+++ frontend/packages/reality-api-client/package.json
@@ -21,5 +21,5 @@
   "peerDependencies": {
     "react": "^18.0.0 || ^19.0.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
