+++ frontend/packages/shared/package.json
@@ -15,5 +15,5 @@
   "devDependencies": {
     "typescript": "^5.3.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
