+++ frontend/packages/ui-kit/package.json
@@ -18,5 +18,5 @@
     "@types/react-dom": "^18.2.0",
     "typescript": "^5.3.0"
   },
-  "version": "0.2.89"
+  "version": "0.2.90"
 }
