+++ mobile-native/gradle.properties
@@ -13,5 +13,5 @@ android.useAndroidX=true
 android.nonTransitiveRClass=true
 
 # App version (synced from VERSION file)
-app.versionName=0.2.89
-app.versionCode=2089
+app.versionName=0.2.90
+app.versionCode=2090
